requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
discord.py[voice]~=1.5.1
ffmpeg
idna==2.9
idna-ssl==1.1.0
multidict==4.7.5
pycparser==2.20
PyNaCl==1.3.0
requests==2.24.0
six==1.15.0
typing-extensions==3.7.4.2
urllib3==1.25.11
websockets==6.0
yarl==1.4.2